proper?
