+++ src/passes/param-utils.cpp
@@ -246,4 +246,27 @@ SortedVector applyConstantValues(const std::vector<Function*>& funcs,
   return optimized;
 }
 
+
+void removeReturns(Function* func, Module* module) {
+  // Remove any return values.
+  struct ReturnUpdater : public PostWalker<ReturnUpdater> {
+    Module* module;
+    ReturnUpdater(Function* func, Module* module) : module(module) {
+      walk(func->body);
+    }
+    void visitReturn(Return* curr) {
+      auto* value = curr->value;
+      assert(value);
+      curr->value = nullptr;
+      Builder builder(*module);
+      replaceCurrent(builder.makeSequence(builder.makeDrop(value), curr));
+    }
+  } returnUpdater(func, module);
+
+  // Remove any value flowing out.
+  if (func->body->type.isConcrete()) {
+    func->body = Builder(*module).makeDrop(func->body);
+  }
+}
+
 } // namespace wasm::ParamUtils
