+++ src/passes/param-utils.h
@@ -92,6 +92,11 @@ SortedVector applyConstantValues(const std::vector<Function*>& funcs,
                                  const std::vector<CallRef*>& callRefs,
                                  Module* module);
 
+// Removes any returns from a function. This does not change the function's type
+// and leaves that part to the caller - it just updates all Return expressions
+// (and a value flowing out) to not return anything. 
+void removeReturns(Function* func, Module* module);
+
 } // namespace wasm::ParamUtils
 
 #endif // wasm_ir_function_h
